added macos support (opengl 3.3 core + fp32)

+++ MandelbrotSet/Shaders/JuliaSet.glsl
@@ -1,5 +1,5 @@
 #shader vertex
-#version 450 core
+#version 330 core
 
 layout(location = 0) in vec3 a_Position;
 
@@ -9,24 +9,24 @@ void main()
 }
 
 #shader fragment
-#version 450 core
+#version 330 core
 layout(location = 0) out vec4 o_Color;
 
-uniform int u_MaxIterations;
-uniform dvec2 u_ScreenSize;
-uniform double u_Zoom;
-uniform dvec2 u_Offset;
-uniform vec4 u_Color;
+uniform int   u_MaxIterations;
+uniform vec2  u_ScreenSize;
+uniform float u_Zoom;
+uniform vec2  u_Offset;
+uniform vec4  u_Color;
 uniform float u_RealComponent;
 uniform float u_ImaginaryComponent;
 
-float JuliaSet(dvec2 c)
+float JuliaSet(vec2 c)
 {
 	int n = 0;
-	dvec2 z = c;
+	vec2 z = c;
 	for (n = 0; n < u_MaxIterations; n++)
 	{
-		dvec2 znew;
+		vec2 znew;
 		znew.x = (z.x * z.x) - (z.y * z.y) + u_RealComponent;
 		znew.y = (2.0 * z.x * z.y) + u_ImaginaryComponent;
 		z = znew;
@@ -47,7 +47,7 @@ vec3 MapToColor(float v)
 
 void main()
 {
-	float pixelValue = JuliaSet(((gl_FragCoord.xy - u_ScreenSize / 2) / u_Zoom) - u_Offset);
+	float pixelValue = JuliaSet(((gl_FragCoord.xy - u_ScreenSize / 2.0) / u_Zoom) - u_Offset);
 	vec3 color = MapToColor(pixelValue);
 	o_Color = vec4(color, 1.0);
 }
